feat: improvement on copy api key ux
