+++ web-portal/frontend/app/dashboard/page.tsx
@@ -3,27 +3,7 @@ import NewAppModal from "./modal.component";
 import { Suspense } from "react";
 import { Container, Flex, Button, Title } from "@mantine/core";
 import Link from "next/link";
-import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
-
-const apiUrl = process.env.API_ENDPOINT || "http://localhost:4000/";
-// TODO: setup a central const and their validation file
-
-async function getTenant() {
-  const id = cookies().get("tenant")?.value;
-
-  if (!id) {
-    redirect("/login");
-  }
-
-  const response = await fetch(`${apiUrl}tenant/${id}`);
-
-  if (!response.ok) {
-    throw new Error("Failed to fetch tenant");
-  }
-
-  return response.json();
-}
+import { getTenant } from "./actions";
 
 export default async function User() {
   const tenant = await getTenant();
@@ -43,13 +23,13 @@ export default async function User() {
             paddingTop: 200,
           }}
         >
-          <NewAppModal id={tenant?.id} />
+          <NewAppModal />
           <Flex
             justify={"space-between"}
             align={"center"}
             style={{ marginBottom: 20 }}
           >
-            <Title order={5}>Tenant id: {tenant?.id}</Title>
+            <Title order={5}>logged in as: {tenant?.id}</Title>
             <Link href="?new=app">
               <Button> New App</Button>
             </Link>
