+++ web-portal/frontend/package.json
@@ -18,6 +18,7 @@
     "@mantine/notifications": "^7.4.2",
     "@mantine/nprogress": "^7.4.2",
     "@mantine/spotlight": "^7.4.2",
+    "@tabler/icons-react": "^2.47.0",
     "dayjs": "^1.11.10",
     "next": "14.0.4",
     "react": "^18",
