+++ web-portal/frontend/utils/consts.ts
@@ -1,0 +1,1 @@
+export const apiUrl = process.env.API_ENDPOINT || "http://localhost:4000/";
